requirements.txt
Flask==3.1.2
google-genai==1.33.0
python-dotenv==1.1.1
requests==2.32.5
